refac: reset vector db
